Move weft into internal
