+++ cmd/fdsn-ws/fdsn_dataselect.go
@@ -4,8 +4,8 @@ import (
 	"bytes"
 	"fmt"
 	"github.com/GeoNet/fdsn/internal/fdsn"
+	"github.com/GeoNet/fdsn/internal/weft"
 	"github.com/GeoNet/kit/mseed"
-	"github.com/GeoNet/weft"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/client"
 	"github.com/aws/aws-sdk-go/aws/session"
