+++ internal/weft/result.go
